refactor(reqgroup): moved to async package
